refactor: Fare 도메인의 패키지 이동
