+++ src/main/java/wooteco/subway/domain/Fare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.line;
+package wooteco.subway.domain;
 
 public class Fare {
