+++ src/test/java/wooteco/subway/domain/FareTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.line;
+package wooteco.subway.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
